# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "AI 聊天助手后端 API",
        "title": "GopherAI API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9090",
    "basePath": "/api/v1",
    "paths": {
        "/AI/chat/send-new-session": {
            "post": {
                "description": "创建新会话并发送用户消息，返回 AI 完整回答",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI 聊天"
                ],
                "summary": "创建会话并发送消息",
                "parameters": [
                    {
                        "description": "用户问题和模型类型",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/session.CreateSessionAndSendMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/session.CreateSessionAndSendMessageResponse"
                        }
                    }
                }
            }
        },
        "/AI/chat/sessions": {
            "get": {
                "description": "获取当前用户的所有历史会话列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI 聊天"
                ],
                "summary": "获取会话列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/session.GetUserSessionsResponse"
                        }
                    }
                }
            }
        },
        "/user/captcha": {
            "post": {
                "description": "向指定邮箱发送验证码，用于注册流程",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "发送邮箱验证码",
                "parameters": [
                    {
                        "description": "邮箱地址",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.CaptchaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.CaptchaResponse"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "通过邮箱、验证码和密码注册账号，注册成功自动返回 JWT Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "注册参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.RegisterResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "code.Code": {
            "type": "integer",
            "format": "int64",
            "enum": [
                1000,
                2001,
                2002,
                2003,
                2004,
                2005,
                2006,
                2007,
                2008,
                2009,
                2010,
                3001,
                4001,
                5001,
                5002,
                5003
            ],
            "x-enum-varnames": [
                "CodeSuccess",
                "CodeInvalidParams",
                "CodeUserExist",
                "CodeUserNotExist",
                "CodeInvalidPassword",
                "CodeNotMatchPassword",
                "CodeInvalidToken",
                "CodeNotLogin",
                "CodeInvalidCaptcha",
                "CodeRecordNotFound",
                "CodeIllegalPassword",
                "CodeForbidden",
                "CodeServerBusy",
                "AIModelNotFind",
                "AIModelCannotOpen",
                "AIModelFail"
            ]
        },
        "model.SessionInfo": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "sessionId": {
                    "type": "string"
                }
            }
        },
        "session.CreateSessionAndSendMessageRequest": {
            "type": "object",
            "required": [
                "modelType",
                "question"
            ],
            "properties": {
                "modelType": {
                    "description": "模型类型;",
                    "type": "string"
                },
                "question": {
                    "description": "用户问题;",
                    "type": "string"
                }
            }
        },
        "session.CreateSessionAndSendMessageResponse": {
            "type": "object",
            "properties": {
                "Information": {
                    "description": "AI回答",
                    "type": "string"
                },
                "sessionId": {
                    "description": "当前会话ID",
                    "type": "string"
                },
                "status_code": {
                    "$ref": "#/definitions/code.Code"
                },
                "status_msg": {
                    "type": "string"
                }
            }
        },
        "session.GetUserSessionsResponse": {
            "type": "object",
            "properties": {
                "sessions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SessionInfo"
                    }
                },
                "status_code": {
                    "$ref": "#/definitions/code.Code"
                },
                "status_msg": {
                    "type": "string"
                }
            }
        },
        "user.CaptchaRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "user.CaptchaResponse": {
            "type": "object",
            "properties": {
                "status_code": {
                    "$ref": "#/definitions/code.Code"
                },
                "status_msg": {
                    "type": "string"
                }
            }
        },
        "user.RegisterRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "captcha": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "user.RegisterResponse": {
            "type": "object",
            "properties": {
                "status_code": {
                    "$ref": "#/definitions/code.Code"
                },
                "status_msg": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        }
    }
}